// src/app/shared/model/music.ts
// export class Music {
//     uid: string;
//     titleName: string;
//     albumName: string;
//     composerName: string;
//     artistName: string;
//     releasedDate: string;
//     albumimage: string
// }
export interface IMusic {
    uid: string;
    titleName: string;
    albumName: string;
    composerName: string;
    artistName: string;
    releasedDate: string;
    albumimage: string
}